Using pad_to_size function to remove deuplicate code in pad_arrays and moving to util
