+++ test/dataset/test_variable_length.py
@@ -17,7 +17,6 @@ from typing import Dict, Any
 # Third-party imports
 import mxnet as mx
 import numpy as np
-import pandas as pd
 import pytest
 
 # First-party imports
@@ -27,7 +26,8 @@ from gluonts.dataset.loader import (
     TrainDataLoader,
     InferenceDataLoader,
 )
-from gluonts.mx.batchify import batchify, stack, _pad_arrays
+from gluonts.mx.batchify import batchify, stack
+from gluonts.testutil.dummy_datasets import get_dataset
 from gluonts.transform import (
     ContinuousTimeInstanceSplitter,
     ContinuousTimeUniformSampler,
@@ -36,37 +36,6 @@ from gluonts.transform import (
 
 @pytest.fixture
 def pp_dataset():
-    def get_dataset():
-
-        data_entry_list = [
-            {
-                "target": np.c_[
-                    np.array([0.2, 0.7, 0.2, 0.5, 0.3, 0.3, 0.2, 0.1]),
-                    np.array([0, 1, 2, 0, 1, 2, 2, 2]),
-                ].T,
-                "start": pd.Timestamp("2011-01-01 00:00:00", freq="H"),
-                "end": pd.Timestamp("2011-01-01 03:00:00", freq="H"),
-            },
-            {
-                "target": np.c_[
-                    np.array([0.2, 0.1, 0.2, 0.5, 0.4]),
-                    np.array([0, 1, 2, 1, 1]),
-                ].T,
-                "start": pd.Timestamp("2011-01-01 00:00:00", freq="H"),
-                "end": pd.Timestamp("2011-01-01 03:00:00", freq="H"),
-            },
-            {
-                "target": np.c_[
-                    np.array([0.2, 0.7, 0.2, 0.5, 0.1, 0.2, 0.1]),
-                    np.array([0, 1, 2, 0, 1, 0, 2]),
-                ].T,
-                "start": pd.Timestamp("2011-01-01 00:00:00", freq="H"),
-                "end": pd.Timestamp("2011-01-01 03:00:00", freq="H"),
-            },
-        ]
-
-        return ListDataset(data_entry_list, freq="H", one_dim_target=False,)
-
     return get_dataset
